First sketch of movign listening and validation/decryption/content out of CoValueCore
